+++ clickhouse_backend/models/__init__.py
@@ -1,3 +1,5 @@
+from .aggregates import *  # noqa: F401,F403
+from .aggregates import __all__ as aggregates_all
 from .base import ClickhouseModel
 from .engines import *  # noqa: F401,F403
 from .engines import __all__ as engines_all
@@ -7,9 +9,6 @@ from .functions import *  # noqa: F401,F403
 from .functions import __all__ as fucntions_all
 from .indexes import *  # noqa: F401,F403
 from .indexes import __all__ as indexes_all
-from .aggregates import *  # noqa: F401,F403
-from .aggregates import __all__ as aggregates_all
-
 
 __all__ = [
     "ClickhouseModel",
