+++ clickhouse_backend/models/functions/random.py
@@ -1,0 +1,9 @@
+from .base import Func
+
+__all__ = [
+    "Rand",
+]
+
+
+class Rand(Func):
+    arity = 0
